[demo] areas == areas, areas != arcs
